fixes attributes appearing in multiple levels of GeoJSON hash, starts to break tests into separate files
